requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
safetensors>=0.4.0
pyyaml>=6.0
